Modified startup sequence for simplification and to reduce errors
