+++ ariac_setup/ariac_setup/ready.py
@@ -1,0 +1,11 @@
+import asyncio
+
+from ariac_setup.utils import ROSAsyncAdapter
+
+
+async def run():
+    await ROSAsyncAdapter.wait_for_gz_clock()
+
+
+def main():
+    asyncio.run(run())
